+++ lib/ArrayExactSize.h
@@ -15,7 +15,7 @@ public:
     std::vector<Domain*> mValues;
 
 public:
-    ExactSize(const Environment &environment, const unsigned size, const Domain* value);
+    ExactSize(const Environment &environment, const uint64_t size, const Domain* value);
     ExactSize(const Environment &environment, const std::vector<Domain*> values);
     ExactSize(const ExactSize &exactSize);
     virtual ~ExactSize();
